requirements.txt
python-dotenv
openai 
pypdf2 
faiss-cpu 
tiktoken
pandas 
numpy 
langchain
matplotlib
plotly
scipy
scikit-learn
streamlit
streamlit_chat
streamlit_jupyter==0.2.1
chromadb 
sentence_transformers
unstructured
pinecone-client
pypdf
duckduckgo-search
autopep8
nest_asyncio
replicate
joblib
nbdev==2.3.12
soundfile
sounddevice
playsound


# wikipedia==1.4.0
# youtube-transcript-api==0.6.0
# google-api-python-client 
# google-auth-httplib2 
# google-auth-oauthlib
# google-search-result
# pytube==12.1.3
# unstructured==0.6.1